Migrate Clean History room action
