--- apps/meteor/app/ui-clean-history/client/index.ts
@@ -1,1 +1,0 @@
-import './lib/tabBar';
