+++ apps/meteor/client/importPackages.ts
@@ -40,7 +40,6 @@ import '../app/slashcommands-topic/client';
 import '../app/slashcommands-unarchiveroom/client';
 import '../app/tokenpass/client';
 import '../app/ui/client';
-import '../app/ui-clean-history/client';
 import '../app/webdav/client';
 import '../app/webrtc/client';
 import '../app/wordpress/client';
